learn_springboot_all/learn_springboot_2/src/main/java/com/example/demo/domain/UserPo.java
package com.example.demo.domain;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class UserPo {

    @Value("${com.name}")
    public String name;
    @Value("${com.age}")
    public Integer age;

    @Value("${com.info}")
    public String info;

    public String getInfo() {
        return info;
    }

    public void setInfo(String info) {
        this.info = info;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Integer getAge() {
        return age;
    }

    public void setAge(Integer age) {
        this.age = age;
    }
}
